requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
decorator==4.4.2
distro==1.9.0
fake-useragent==2.0.3
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
lxml==5.3.0
lxml_html_clean==0.4.1
meta_ai_api==1.2.5
moviepy==1.0.3
numpy==2.2.2
oauthlib==3.2.2
openai==1.60.1
parse==1.20.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
pyparsing==3.2.1
pyppeteer==0.0.25
pyquery==2.0.1
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.15.1
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
w3lib==2.3.1
websockets==14.2
